rename: 잘못된 폴더 구조 변경game/hooks/useCoach -> player/hooks/useCoach (#99)
